platforms/platform-01/eco-backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Validation and Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Task Queue
celery==5.3.6
flower==2.0.1

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
structlog==24.1.0

# Rate Limiting and Circuit Breaker
slowapi==0.1.9
circuitbreaker==1.4.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Utilities
tenacity==8.2.3
python-json-logger==2.0.7
orjson==3.9.10
